clean up affordance code
